feat(temporal102): init age-estimation sample
so it turns out that the API used as part of this sample doesn't support
TLS1.2+EMS (Extended Main Secret), which 'http-client-tls' is set up to
require by default.

there are a few ways we could have handled this, but I think it actually
makes for a nice example for how dependency injection works for the
activity's 'MonadReader' instance.